fix int overflow

+++ include/AGO/algorithm/gene_type_analyzer_sqalign.hpp
@@ -304,7 +304,7 @@ class GeneTypeAnalyzerSqalign
         {
             if( rnafolds.find( mir.first ) == rnafolds.end() ) continue;
 
-            start = std::stoi( mir.second.strand == '+' ? rnafolds[ mir.first ][1] : rnafolds[ mir.first ][2] ) + 1;
+            start = std::stoll( mir.second.strand == '+' ? rnafolds[ mir.first ][1] : rnafolds[ mir.first ][2] ) + 1;
             shift = start > mir.second.refstart ? ( start - mir.second.refstart ) : ( mir.second.refstart - start );
 
             for( auto& seq : mir.second.reads_vec )
